Change `parse()` function's arguments`
Input to `parse()` used to be `yaml_input` for the actual YAML, and `file_in` for the YAML file path.
THe latter was only used to resolve relative image paths, not for reading the actual file that was passed.
Therefore, `file_in` was changed to `base_path`; this argument receives the *directory* of the original YAML file, which is more intuitive for the actual use case.