Avoid a node size greater than the HTML label
Bug: When the HTML label is narrow (e.g. connectors in tutorial01),
then the white node background is wider, and when the HTML label is
low (e.g. ferrules in tutorial05 and totorial06), then the white
node background is taller. Such errors are easily seen with a
non-white bgcolor.

Fix: Set node width, height, and margin to zero to let the actual
size of the node be entirely determined by the label.